+++ db/migrations/2_create_account_tables.down.sql
@@ -1,0 +1,5 @@
+DROP TABLE IF EXISTS loans;
+DROP TABLE IF EXISTS deposits;
+DROP TABLE IF EXISTS accounts;
+DROP TYPE account_type;
+DROP TABLE IF EXISTS currencies;