+++ internal/repo/error.go
@@ -5,4 +5,8 @@ import "errors"
 var (
 	ErrUserAlreadyExists = errors.New("user already exists")
 	ErrUserNotFound      = errors.New("user doesn't exists")
+
+	ErrCurrencyNotFound = errors.New("currency doesn't exists")
+
+	ErrAccountNotFound = errors.New("account doesn't exists")
 )