+++ CHANGELOG.md
@@ -3,5 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Operations for accounts.
 
 [unreleased]: https://github.com/lotostudio/financial-api/tree/develop